+++ source/DSCResources/SqlServer/SqlServer.schema.psm1
@@ -107,7 +107,7 @@ configuration SqlServer
     Import-DscResource -ModuleName SecurityPolicyDsc -ModuleVersion 2.10.0.0
     . "$resourcePath\Windows.SecurityOption.ps1"
 
-    Import-DscResource -ModuleName AccessControlDsc -ModuleVersion 1.4.1
+    Import-DscResource -ModuleName AccessControlDsc -ModuleVersion 1.4.2
     . "$resourcePath\Windows.AccessControl.ps1"
 
     Import-DscResource -ModuleName PSDSCresources -ModuleVersion 2.12.0.0